docker/requirements.txt
fastapi
uvicorn
pillow
backgroundremover
openai==0.28
supabase
python-dotenv